updated ebp for bundle of 2
