+++ dashboard/src/components/bundles2/Bundle19.jsx
@@ -1,16 +1,163 @@
-import { Box } from "@mui/material";
-import Header from "../Header";
 import LineChart from "../LineChart";
+import { Box, IconButton, Typography, useTheme } from "@mui/material";
+import DownloadOutlinedIcon from "@mui/icons-material/DownloadOutlined";
+import EmailIcon from "@mui/icons-material/Email";
+import PointOfSaleIcon from "@mui/icons-material/PointOfSale";
+import PersonAddIcon from "@mui/icons-material/PersonAdd";
+import TrafficIcon from "@mui/icons-material/Traffic";
+import Header from "../Header";
+import StatBox from "../StatBox";
+import { tokens } from "../../theme";
+import BundleBar from "../../scenes/ebp/BundleBar";
 
 const Bundle19 = () => {
+  const theme = useTheme();
+  const colors = tokens(theme.palette.mode);
+
   return (
     <Box m="20px">
-      <Header title="Bundle: Wings Of Time & Sea Aquarium" subtitle="Revenue of Bundle" />
-      <Box height="75vh">
-        <LineChart />
+      {/* HEADER */}
+      <Box display="flex" justifyContent="space-between" alignItems="center">
+        <Header title="Explore Bundle Prices" subtitle="Bundle Prices" />
+
       </Box>
+
+      <BundleBar />
+
+    <Box m="20px">
+      <Header title="Bundle: Wings Of Time & Sea Aquarium" subtitle="Revenue of Bundle" />
+
+          {/* GRID & CHARTS */}
+    <Box
+      display="grid"
+      gridTemplateColumns="repeat(12, 1fr)"
+      gridAutoRows="140px"
+      gap="20px"
+    >
+      {/* ROW 1 */}
+      <Box
+        gridColumn="span 3"
+        backgroundColor={colors.primary[400]}
+        display="flex"
+        alignItems="center"
+        justifyContent="center"
+      >
+        <StatBox
+          title="12,361"
+          subtitle="Emails Sent"
+          progress="0.75"
+          increase="+14%"
+          icon={
+            <EmailIcon
+              sx={{ color: colors.greenAccent[600], fontSize: "26px" }}
+            />
+          }
+        />
+      </Box>
+      <Box
+        gridColumn="span 3"
+        backgroundColor={colors.primary[400]}
+        display="flex"
+        alignItems="center"
+        justifyContent="center"
+      >
+        <StatBox
+          title="431,225"
+          subtitle="Sales Obtained"
+          progress="0.50"
+          increase="+21%"
+          icon={
+            <PointOfSaleIcon
+              sx={{ color: colors.greenAccent[600], fontSize: "26px" }}
+            />
+          }
+        />
+      </Box>
+      <Box
+        gridColumn="span 3"
+        backgroundColor={colors.primary[400]}
+        display="flex"
+        alignItems="center"
+        justifyContent="center"
+      >
+        <StatBox
+          title="32,441"
+          subtitle="New Clients"
+          progress="0.30"
+          increase="+5%"
+          icon={
+            <PersonAddIcon
+              sx={{ color: colors.greenAccent[600], fontSize: "26px" }}
+            />
+          }
+        />
+      </Box>
+      <Box
+        gridColumn="span 3"
+        backgroundColor={colors.primary[400]}
+        display="flex"
+        alignItems="center"
+        justifyContent="center"
+      >
+        <StatBox
+          title="1,325,134"
+          subtitle="Traffic Received"
+          progress="0.80"
+          increase="+43%"
+          icon={
+            <TrafficIcon
+              sx={{ color: colors.greenAccent[600], fontSize: "26px" }}
+            />
+          }
+        />
+      </Box>
+
+      {/* ROW 2 */}
+      <Box
+        gridColumn="span 8"
+        gridRow="span 2"
+        backgroundColor={colors.primary[400]}
+      >
+        <Box
+          mt="25px"
+          p="0 30px"
+          display="flex "
+          justifyContent="space-between"
+          alignItems="center"
+        >
+          <Box>
+            <Typography
+              variant="h5"
+              fontWeight="600"
+              color={colors.grey[100]}
+            >
+              Revenue Generated
+            </Typography>
+            <Typography
+              variant="h3"
+              fontWeight="bold"
+              color={colors.greenAccent[500]}
+            >
+              $59,342.32
+            </Typography>
+          </Box>
+          <Box>
+            <IconButton>
+              <DownloadOutlinedIcon
+                sx={{ fontSize: "26px", color: colors.greenAccent[500] }}
+              />
+            </IconButton>
+          </Box>
+        </Box>
+        <Box height="250px" m="-20px 0 0 0">
+          <LineChart isDashboard={true} />
+        </Box>
+      </Box>
+      </Box>
+    </Box>
     </Box>
   );
 };
 
 export default Bundle19;
+
